+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/quartiq/idsp/compare/v0.14.1..main) - 2024-01-15
+
+### Added
+
+* `Dsm`: Delta sigma modulator/noise shaper in MASH-(1)^K architecture.
+
 ## [0.14.1](https://github.com/quartiq/idsp/compare/v0.14.0..v0.14.1) - 2024-01-15
 
 * Fixed changelog